requirements.txt
Django==1.7.1
MySQL-python==1.2.5
Pillow==2.6.1
PyMySQL==0.6.3
South==1.0.1
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-angular==0.7.10
django-guardian==1.2.4
django-less==0.7.2
django-messages==0.5.1
django-pipeline==1.4.2
django-sitetree==1.2.0
django-toolbelt==0.0.1
django-twitter-bootstrap==3.3.0
django-userena==1.3.1
easy-thumbnails==2.2
futures==2.2.0
gunicorn==19.1.1
html2text==2014.9.25
psycopg2==2.5.4
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
yolk==0.4.3
